// HealthCare/src/Interfaces/SampleAccession/patient-info-response.ts
export interface PatientInfoResponse {
    sampleId: number;
    patientName: string;
    referDoctor: string;
    referLab: string;
    centerCode: string;
    workOrderStatus: string;
    patientCode: string;
    emailId: string;
    paymentType: string;
    sampleType: string;
    barcode: string;
    allowAccession: string;
    partnerId: string;
    testName: string;
}